requirements.txt
absl-py==1.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
filelock==3.8.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
more-itertools==8.14.0
numpy==1.21.6
packaging==21.3
pyparsing==3.0.9
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
tokenizers==0.12.1
torch==1.12.1
tqdm==4.64.0
transformers==4.21.1
typing_extensions==4.3.0
urllib3==1.26.11
zipp==3.8.1
